requirements.txt
requests
google-cloud-texttospeech
speechrecognition
pyaudio